t-whatpm/swml/forms-specific-1.dat

#data
[[#comment]]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <sw form>
|       ref="comment"

#data
a[[#comment]]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "a"
|       <sw form>
|         ref="comment"

#data
[[#comment:arg]]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <sw form>
|       parameter="arg"
|       ref="comment"

#data
a[[#comment:arg]]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "a"
|       <sw form>
|         parameter="arg"
|         ref="comment"
